@@ -3,22 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fontsource/roboto": "^4.5.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@types/node": "^12.20.21",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.18",
+    "@emotion/react": "^11.6.0",
+    "@emotion/styled": "^11.6.0",
+    "@fontsource/roboto": "^4.5.1",
+    "@mui/material": "^5.1.1",
+    "@mui/styles": "^5.1.1",
+    "@reduxjs/toolkit": "^1.6.2",
+    "@types/node": "^16.11.7",
+    "@types/react": "^17.0.35",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-redux": "^7.1.20",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-google-charts": "^3.0.15",
-    "react-redux": "^7.2.5",
-    "redux": "^4.1.1",
-    "socket.io-client": "^4.2.0",
-    "typescript": "^4.4.2",
-    "web-vitals": "^1.1.2"
+    "react-redux": "^7.2.6",
+    "redux": "^4.1.2",
+    "socket.io-client": "^4.3.2",
+    "typescript": "^4.5.2",
+    "web-vitals": "^2.1.2"
   },
   "scripts": {
     "start": "react-scripts --openssl-legacy-provider start",
@@ -69,16 +70,16 @@
     ]
   },
   "devDependencies": {
-    "@cypress/react": "^5.9.4",
-    "@cypress/webpack-dev-server": "^1.5.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "react-scripts": "4.0.3",
-    "cypress": "^8.3.1",
-    "husky": "^7.0.2",
-    "lint-staged": "^11.1.2",
-    "prettier": "^2.3.2"
+    "@cypress/react": "^5.10.3",
+    "@cypress/webpack-dev-server": "^1.7.0",
+    "@testing-library/jest-dom": "^5.15.0",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.0.2",
+    "cypress": "^8.7.0",
+    "husky": "^7.0.4",
+    "lint-staged": "^12.0.2",
+    "prettier": "^2.4.1",
+    "react-scripts": "4.0.3"
   }
 }
